UI-related documentation: reviewed and adapted
Reviewed and adapted the UI documentation by @tonysani
The overall structure has changed, certain information was removed from the "Tutorials" section and moved to other topics.
Some of the original topics are modified to better suit the final product.

Some issues are still present in this commit. I will discuss them with @Mingela and aim to fix them in the next commit.